WIP: Moment display at the end of the turn only
